Customization made to the cx_Freeze scripts.
ModuleFinder.py has a fix that gets around several limitations of the original when it comes to
determining dependencies.
ConsoleSetLibPath.py has been modified to set the LD_LIBRARY_PATH var to the
'dabolibs' directory off the main distro directory.
build.sh is the shell script used to generate the frozen distro. It will
require some pathing customziation in order to work on whatever machine it
is run on.