Added a buildwin script because I'm too lazy to remember what to send to setup.py.
Added a sys._daborun flag so the framework knows how to find out what
HomeDirectory should be.

Inserted the path to the application file instead of appending it, as
it really should come first in the search path, not last, which is how
non-frozen apps behave.

I'm committing this now from Linux so I can test on Windows.